+++ web-app/server.js
@@ -4,6 +4,8 @@ const cors = require("cors");
 
 const apiRouter = require("./router");
 const maindataRouter = require("./server-routes/maindata-router");
+const popularRouter = require("./server-routes/popular-router");
+const ordersRouter = require("./server-routes/orders-router");
 
 const app = express();
 const port = 5000;
@@ -23,8 +25,10 @@ app.use(express.json());
 app.use(express.static(path.join(__dirname, "client/build")));
 
 // Bereitstellen der Router
-app.use("/api", apiRouter);
+//app.use("/api", apiRouter);
 app.use("/api/maindata", maindataRouter);
+app.use("/api/popular", popularRouter);
+app.use("/api", ordersRouter);
 
 // Behandelt alle Anfragen, die nicht mit den oben genannten übereinstimmen
 app.get("*", (req, res) => {
